After CR: Reorganized plugin structure #94
